Flatten out dashboard header directory
- includes removing some duped code relating to the domains picker, which I replaced with importing the same stuff from the collection picker code for now
