+++ src/dashboard-refactor/header/index.tsx
@@ -6,11 +6,13 @@ import { SETTINGS_URL } from 'src/constants'
 import SearchBar, { SearchBarProps } from './search-bar'
 import { SyncStatusIconState } from './types'
 import SyncStatusMenu, { SyncStatusMenuProps } from './sync-status-menu'
-import SidebarHeader, { SidebarHeaderProps } from './sidebar-header'
 import styles, { fonts } from '../styles'
 import { Icon } from '../styled-components'
 import Margin from '../components/Margin'
 import { sizeConstants } from 'src/dashboard-refactor/constants'
+import SidebarToggle from './sidebar-toggle'
+import { SidebarLockedState } from '../lists-sidebar/types'
+import { HoverState } from '../types'
 
 const Container = styled.div`
     height: ${sizeConstants.header.heightPx}px;
@@ -94,8 +96,37 @@ const SyncStatusHeaderText = styled.span<{
     ${(props) => (props.textCentered ? 'text-align: center;' : '')}
 `
 
+const SidebarHeaderContainer = styled.div`
+    height: 100%;
+    width: ${sizeConstants.listsSidebar.widthPx}px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: flex-start;
+    flex: 1;
+
+    & div {
+        justify-content: flex-start;
+    }
+`
+
+const CollectionTitle = styled.p`
+    margin: 0;
+    font-family: ${fonts.primary.name};
+    font-weight: ${fonts.primary.weight.bold};
+    line-height: 21px;
+    width: 100%;
+    display: flex;
+`
+
+const SidebarToggleBox = styled(Margin)`
+    width: min-content;
+`
+
 export interface HeaderProps {
-    sidebarHeaderProps: SidebarHeaderProps
+    sidebarLockedState: SidebarLockedState
+    sidebarToggleHoverState: HoverState
+    selectedListName?: string
     searchBarProps: SearchBarProps
     syncStatusMenuProps: SyncStatusMenuProps
     syncStatusIconState: SyncStatusIconState
@@ -106,14 +137,24 @@ export default class Header extends PureComponent<HeaderProps> {
 
     render() {
         const {
-            sidebarHeaderProps,
             searchBarProps,
             syncStatusIconState,
             syncStatusMenuProps,
+            selectedListName: selectedList,
         } = this.props
         return (
             <Container>
-                <SidebarHeader {...sidebarHeaderProps} />
+                <SidebarHeaderContainer>
+                    <SidebarToggleBox horizontal="12px">
+                        <SidebarToggle
+                            sidebarLockedState={this.props.sidebarLockedState}
+                            hoverState={this.props.sidebarToggleHoverState}
+                        />
+                    </SidebarToggleBox>
+                    {selectedList && (
+                        <CollectionTitle>{selectedList}</CollectionTitle>
+                    )}
+                </SidebarHeaderContainer>
                 <SearchSection vertical="auto" horizontal="17px">
                     <SearchBar {...searchBarProps} />
                 </SearchSection>
