--- src/dev/stories/dashboard-refactor/header/sidebar-header.tsx
@@ -1,62 +1,0 @@
-import React from 'react'
-import { storiesOf } from '@storybook/react'
-
-import { sidebarToggleProps } from './sidebar-toggle'
-
-import SidebarHeader, {
-    SidebarHeaderProps,
-} from 'src/dashboard-refactor/header/sidebar-header/'
-
-const stories = storiesOf('Dashboard Refactor|Header/Sidebar Header', module)
-
-const { lockedHover, noHover } = sidebarToggleProps
-const sidebarTogglePropsRenamed = {
-    lockedHover: {
-        ...lockedHover,
-        sidebarToggleHoverState: lockedHover.hoverState,
-    },
-    noHover: {
-        ...noHover,
-        sidebarToggleHoverState: noHover.hoverState,
-    },
-}
-
-export const sidebarHeaderPropsTemplate: {
-    open: SidebarHeaderProps
-    closed: SidebarHeaderProps
-    closedSelected: SidebarHeaderProps
-} = {
-    open: {
-        sidebarPeekState: {
-            setSidebarPeekState: () => () => {},
-            isSidebarPeeking: false,
-        },
-        selectedListName: 'Inbox',
-        ...sidebarTogglePropsRenamed.lockedHover,
-    },
-    closed: {
-        sidebarPeekState: {
-            setSidebarPeekState: () => () => {},
-            isSidebarPeeking: false,
-        },
-        ...sidebarTogglePropsRenamed.noHover,
-    },
-    closedSelected: {
-        ...sidebarTogglePropsRenamed.noHover,
-        sidebarPeekState: {
-            setSidebarPeekState: () => () => {},
-            isSidebarPeeking: false,
-        },
-        selectedListName: 'Inbox',
-    },
-}
-
-stories.add('Sidebar Header/Open', () => (
-    <SidebarHeader {...sidebarHeaderPropsTemplate.open} />
-))
-stories.add('Sidebar Header/Closed', () => (
-    <SidebarHeader {...sidebarHeaderPropsTemplate.closed} />
-))
-stories.add('Sidebar Header/Closed with Inbox selected', () => (
-    <SidebarHeader {...sidebarHeaderPropsTemplate.closedSelected} />
-))
